+++ src/Joomlatools/Console/Command/Site/Download.php
@@ -60,7 +60,7 @@ class Download extends AbstractSite
                 'repo',
                 null,
                 InputOption::VALUE_REQUIRED,
-                'Alternative Git repository to clone. To use joomlatools/joomla-platform, use --repo=platform.'
+                'Alternative Git repository to clone. To use joomlatools/platform, use --repo=platform.'
             )
         ;
     }